requirements.txt
python-dotenv==0.19.2
lxml==4.6.3
requests==2.25.1
mecab-python3==1.0.3